// mainServer/src/modules/core/tools/browserAutomationTool/browser-automation.types.ts
export enum BrowserOperation {
  GOTO = 'goto',
  CLICK = 'click',
  FILL = 'fill',
  SCREENSHOT = 'screenshot',
  EVALUATE = 'evaluate',
  PAUSE_FOR_INPUT = 'pause_for_input',
  CLOSE = 'close',
  LIST_TABS = 'list_tabs',
  SWITCH_TAB = 'switch_tab',
  NEW_TAB = 'new_tab',
  WAIT_FOR = 'wait_for',
}